+++ cli/vitest.config.ts
@@ -5,6 +5,8 @@ export default defineConfig({
   test: {
     testTimeout: 20_000,
     hookTimeout: 20_000,
+    setupFiles: ['./tests/test-setup.ts'],
+    disableConsoleIntercept: true,
   },
 })